LayoutTests/media/modern-media-controls/playback-support/playback-support-button-click.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<video src="../../content/test.mp4" style="width: 320px; height: 240px;"></video>
<div id="shadow"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>PlaybackSupport</code> behavior by clicking on the play/pause button.");

const shadowRoot = document.querySelector("div#shadow").attachShadow({ mode: "open" });
const media = document.querySelector("video");
const mediaController = createControls(shadowRoot, media, null);

shouldBeFalse("mediaController.controls.playPauseButton.playing");

media.addEventListener("play", function() {
    debug("Media is playing");
    shouldBeTrue("mediaController.controls.playPauseButton.playing");

    clickPlayPauseButton();
});

media.addEventListener("pause", function() {
    debug("Media is paused");
    shouldBeFalse("mediaController.controls.playPauseButton.playing");
    debug("");
    shadowRoot.host.remove();
    media.remove();
    finishJSTest();
});

function clickPlayPauseButton()
{
    debug("");
    debug("Dispatching click event to the play/pause button");
    mediaController.controls.playPauseButton.element.click();
}

clickPlayPauseButton();

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
